* keyedit.c (keyedit_menu, menu_backsign): New "backsign" command to add 0x19 backsigs to old keys that don't have them.
* misc.c (parse_options): Fix build warning.

* main.h, keygen.c (make_backsig): Make public.

+++ g10/misc.c
@@ -979,7 +979,7 @@ parse_options(char *str,unsigned int *options,
       for(i=0;opts[i].name;i++)
         if(opts[i].help)
 	  printf("%s%*s%s\n",opts[i].name,
-		 maxlen+2-strlen(opts[i].name),"",_(opts[i].help));
+		 maxlen+2-(int)strlen(opts[i].name),"",_(opts[i].help));
 
       g10_exit(0);
     }
